+++ dbt/adapters/databricks/connections.py
@@ -85,10 +85,8 @@ class DatabricksCredentials(Credentials):
 
     def __post_init__(self) -> None:
         if "." in self.schema:
-            logger.warning(
-                f"The specified schema '{self.schema}' contains '.', "
-                "which could cause unexpected behavior.\n"
-                "It will not be allowed in the future release.\n"
+            raise dbt.exceptions.ValidationException(
+                f"The schema should not contain '.': {self.schema}\n"
                 "If you are trying to set a catalog, please use `catalog` instead.\n"
             )